Move Entity Modal Hooks into `/hooks/` Folder js_only
