+++ assets/src/domain/eventEditor/ui/datetimes/hooks/useAddDatetimeModal.ts
@@ -2,35 +2,35 @@ import { useCallback } from 'react';
 import { __ } from '@wordpress/i18n';
 import { FormProps } from 'react-final-form';
 
-import useTicketFormConfig from './ticketForm/useTicketFormConfig';
+import useDateFormConfig from '../dateForm/useDateFormConfig';
 import { useFormModal, FormModal, ModalClose } from '@appLayout/formModal';
-import { useTicketMutator, CreateTicketInput } from '@edtrServices/apollo/mutations';
+import { useDatetimeMutator, CreateDatetimeInput } from '@edtrServices/apollo/mutations';
 import useEvent from '@edtrServices/apollo/queries/events/useEvent';
 
-const useAddTicketModal: FormModal = () => {
-	const { createEntity } = useTicketMutator();
+const useAddDatetimeModal: FormModal = () => {
+	const { createEntity } = useDatetimeMutator();
 	const { closeEditor } = useFormModal();
 
 	const { name: eventName = '' } = useEvent() || {};
 
 	const onClose = useCallback<ModalClose>((): void => {
-		closeEditor('addTicket');
+		closeEditor('addDatetime');
 	}, [closeEditor]);
 
-	const onSubmit = useCallback<FormProps<CreateTicketInput>['onSubmit']>(
+	const onSubmit = useCallback<FormProps<CreateDatetimeInput>['onSubmit']>(
 		(fields): void => {
 			createEntity(fields);
 		},
 		[createEntity]
 	);
 
-	const formConfig = useTicketFormConfig(null, { onSubmit });
+	const formConfig = useDateFormConfig(null, { onSubmit });
 
 	return {
-		title: `${eventName}: ${__('New Ticket Details')}`,
+		title: `${eventName}: ${__('New Datetime')}`,
 		formConfig,
 		onClose,
 	};
 };
 
-export default useAddTicketModal;
+export default useAddDatetimeModal;
