+++ assets/src/domain/eventEditor/ui/tickets/hooks/useAddTicketModal.ts
@@ -2,35 +2,35 @@ import { useCallback } from 'react';
 import { __ } from '@wordpress/i18n';
 import { FormProps } from 'react-final-form';
 
-import useDateFormConfig from './dateForm/useDateFormConfig';
+import useTicketFormConfig from '../ticketForm/useTicketFormConfig';
 import { useFormModal, FormModal, ModalClose } from '@appLayout/formModal';
-import { useDatetimeMutator, CreateDatetimeInput } from '@edtrServices/apollo/mutations';
+import { useTicketMutator, CreateTicketInput } from '@edtrServices/apollo/mutations';
 import useEvent from '@edtrServices/apollo/queries/events/useEvent';
 
-const useAddDatetimeModal: FormModal = () => {
-	const { createEntity } = useDatetimeMutator();
+const useAddTicketModal: FormModal = () => {
+	const { createEntity } = useTicketMutator();
 	const { closeEditor } = useFormModal();
 
 	const { name: eventName = '' } = useEvent() || {};
 
 	const onClose = useCallback<ModalClose>((): void => {
-		closeEditor('addDatetime');
+		closeEditor('addTicket');
 	}, [closeEditor]);
 
-	const onSubmit = useCallback<FormProps<CreateDatetimeInput>['onSubmit']>(
+	const onSubmit = useCallback<FormProps<CreateTicketInput>['onSubmit']>(
 		(fields): void => {
 			createEntity(fields);
 		},
 		[createEntity]
 	);
 
-	const formConfig = useDateFormConfig(null, { onSubmit });
+	const formConfig = useTicketFormConfig(null, { onSubmit });
 
 	return {
-		title: `${eventName}: ${__('New Datetime')}`,
+		title: `${eventName}: ${__('New Ticket Details')}`,
 		formConfig,
 		onClose,
 	};
 };
 
-export default useAddDatetimeModal;
+export default useAddTicketModal;
